% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{markov_mix_ex}
\alias{markov_mix_ex}
\title{Mixture of Markov chain example}
\format{
A \code{\link{MarkovMix}} object.
}
\usage{
markov_mix_ex
}
\description{
A mixture of 2-order Markov chain fit from 100 random sequences
with 4 states
(A, B, C, D)
and 3 components.
}
\note{
Change log:
\itemize{
\item{0.1.0 Xiurui Zhu - Initiate the dataset.}
}
}
\author{
Xiurui Zhu
}
\keyword{datasets}
