% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markovmix-package.R
\name{markovmix-package}
\alias{markovmix-package}
\title{markovmix package}
\description{
\code{markovmix} package fits mixture of Markov chains of higher orders from multiple sequences.
It is also compatible with ordinary 1-component, 1-order or single-sequence Markov chains.
In addition to the fit function, custom print and predict functions are also provided.
}
\note{
Change log:
\itemize{
\item{0.1.0 Xiurui Zhu - Initiate the package.}
}
}
\author{
Xiurui Zhu
}
